# swagger-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smoothies API",
    "version": "1.0.0",
    "description": "REST API for managing smoothies built with Express and Supabase"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Smoothie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "6f845fbd-c325-4fef-8f53-2dfe08fda109"
          },
          "title": {
            "type": "string",
            "example": "Green Energy Boost"
          },
          "method": {
            "type": "string",
            "example": "Blend all ingredients until smooth."
          },
          "rating": {
            "type": "number",
            "example": 4.5
          },
          "user_id": {
            "type": "string",
            "example": "b112a330-67fb-4d59-8d43-ccfe81f5d11e"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-13T08:00:00Z"
          }
        }
      },
      "SmoothieInput": {
        "type": "object",
        "required": [
          "title",
          "method",
          "rating"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Strawberry Banana Shake"
          },
          "method": {
            "type": "string",
            "example": "Add all ingredients to blender. Mix well."
          },
          "rating": {
            "type": "number",
            "example": 4
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "firstName"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "securepassword123"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "firstName": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "summary": "Login a user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "securepassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "access_token": {
                      "type": "string",
                      "example": "your-jwt-token-here"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "firstName": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          }
        }
      }
    },
    "/auth/signout": {
      "post": {
        "summary": "Log out current user",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get the currently authenticated user",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User data or null",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "oneOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "email": {
                              "type": "string"
                            },
                            "firstName": {
                              "type": "string"
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/smoothies/public": {
      "get": {
        "summary": "Get all smoothies",
        "tags": [
          "Smoothies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of smoothies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalCount": {
                      "type": "integer",
                      "example": 25
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 10
                    },
                    "smoothies": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Smoothie"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/smoothies/my": {
      "get": {
        "summary": "Get smoothies created by the logged-in user",
        "tags": [
          "Smoothies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of user smoothies"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/smoothies/{id}": {
      "get": {
        "summary": "Get a smoothie by ID",
        "tags": [
          "Smoothies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Smoothie found"
          },
          "404": {
            "description": "Smoothie not found"
          }
        }
      },
      "put": {
        "summary": "Update a smoothie",
        "tags": [
          "Smoothies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "method",
                  "rating"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "method": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Smoothie updated"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Smoothie not found or not updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a smoothie",
        "tags": [
          "Smoothies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Smoothie deleted successfully"
          },
          "404": {
            "description": "Smoothie not found"
          }
        }
      }
    },
    "/smoothies": {
      "post": {
        "summary": "Create a new smoothie",
        "tags": [
          "Smoothies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmoothieInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Smoothie created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Smoothie"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}